INTYGFV-16492: Made requested changes to ElementRule and updated error management in Validation converter logic.

+++ certificate-service/domain/src/main/java/se/inera/intyg/certificateservice/domain/certificatemodel/model/RuleLimit.java
@@ -1,5 +1,5 @@
 package se.inera.intyg.certificateservice.domain.certificatemodel.model;
 
-public record RuleLimit(Integer value) implements Rule {
+public record RuleLimit(short value) implements Rule {
 
 }
